Split item handlers into individual files
