+++ internal/pkg/keybindings/item_left.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import "github.com/awesome-gocui/gocui"
+
+type ItemLeftHandler struct {
+	ItemHandler
+	EditModeEnabled *bool
+}
+
+func NewItemLeftHandler(editModeEnabled *bool) *ItemLeftHandler {
+	handler := &ItemLeftHandler{
+		EditModeEnabled: editModeEnabled,
+	}
+	handler.id = HandlerIDItemLeft
+	return handler
+}
+
+func (h ItemLeftHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		tmp := false // memory leak?
+		h.EditModeEnabled = &tmp
+		g.Cursor = false
+		g.SetCurrentView("listWidget")
+		return nil
+	}
+}
